requirements.txt
pyrogram
tgcrypto
python-dotenv
python-telegram-bot
APScheduler
